+++ src/components/App.jsx
@@ -1,4 +1,4 @@
-import { useState } from 'react';
+import { useState, useRef } from 'react';
 import Header from './Header';
 import Contact from './Contact';
 import Education from './Education';
@@ -6,63 +6,54 @@ import Skills from './Skills';
 import Work from './Work';
 import References from './References';
 import Button from './Button';
-import formValidation, { validateInput } from '../utilities/formValidation';
+import validateForms from '../utilities/formValidation';
+import Modal from './Modal';
+import Instructions from './Instructions';
 import '../styles/app.css';
 import '../styles/form.css';
 
 export default function App() {
   const [isEditing, setIsEditing] = useState(false);
+  const [modal, setModal] = useState(false);
+  const deleteRef = useRef();
   const editHandler = () => {
     setIsEditing(!isEditing);
   };
 
-  const validateForms = () => {
-    console.log('check firing!');
-    if (areFormsValid() && isEditing) {
-      setIsEditing(!isEditing);
-    }
-  };
+  // if (isEditing) validateForms(setIsEditing);
 
-  const validateForm = formValidation;
   const buttonProps = isEditing
     ? { type: 'submit', className: 'cv-submit', text: 'Submit CV' }
-    : { type: 'button', className: 'cv-edit', text: 'Edit CV' };
+    : { className: 'cv-edit', text: 'Edit CV' };
 
   return (
     <div className="app">
-      <Button {...buttonProps} onClick={!isEditing ? editHandler : validateForms} />
-      <Header isEditing={isEditing} validateForm={isEditing && validateForm} />
-      <Contact
-        isEditing={isEditing}
-        validateForm={isEditing && validateForm}
-        validateInput={isEditing && validateInput}
+      <Button
+        {...buttonProps}
+        onClick={
+          !isEditing
+            ? editHandler
+            : () => {
+                validateForms(setIsEditing);
+              }
+        }
       />
-      <Education
-        isEditing={isEditing}
-        validateForm={isEditing && validateForm}
-        validateInput={isEditing && validateInput}
-      />
-      <Skills
-        isEditing={isEditing}
-        validateForm={isEditing && validateForm}
-        validateInput={isEditing && validateInput}
-      />
-      <Work
-        isEditing={isEditing}
-        validateForm={isEditing && validateForm}
-        validateInput={isEditing && validateInput}
-      />
-      {/* <References isEditing={isEditing} validateForm={isEditing && validateForm} /> */}
+      <Header isEditing={isEditing} />
+      <Contact isEditing={isEditing} />
+      <Education isEditing={isEditing} setModal={setModal} deleteRef={deleteRef} />
+      <Skills isEditing={isEditing} setModal={setModal} deleteRef={deleteRef} />
+      <Work isEditing={isEditing} setModal={setModal} deleteRef={deleteRef} />
+      <References isEditing={isEditing} setModal={setModal} deleteRef={deleteRef} />
+      {isEditing && (
+        <>
+          <Instructions></Instructions>
+          <Modal
+            openModal={modal}
+            closeModal={() => setModal(false)}
+            {...deleteRef.current}
+          ></Modal>
+        </>
+      )}
     </div>
   );
 }
-
-const areFormsValid = () => {
-  const forms = document.querySelectorAll('form');
-
-  console.log(forms);
-
-  // Return true if all form inputs are valid
-  // Return false if a form input is invalid
-  return true;
-};