+++ src/components/Skills.jsx
@@ -2,18 +2,20 @@ import { useState } from 'react';
 import { skills as initialSkillsState } from '../data/data.initialStates';
 import FormItem from './FormItem';
 import Button from './Button';
+import { validateForm } from '../utilities/formValidation';
 import '../styles/skills.css';
 // https://codesandbox.io/p/sandbox/react-dev-64n8l5?file=%2Fsrc%2FTaskList.js&utm_medium=sandpack
 // https://www.w3.org/WAI/tutorials/forms/labels/
 // https://dev.to/ajones_codes/a-better-guide-to-forms-in-react-47f0
 let nextId = 1;
 
-export default function Skills({ isEditing, validateForm }) {
+export default function Skills({ isEditing, setModal, deleteRef }) {
   const [skillsData, setSkillsData] = useState({
     ...initialSkillsState,
   });
 
   const onChangeHandler = (e) => {
+    // Temporarily does not work
     const input = e.currentTarget;
     const value = input.value;
     const { id, key, prop } = { prop: input.name, ...input.dataset };
@@ -38,8 +40,7 @@ export default function Skills({ isEditing, validateForm }) {
     });
   };
 
-  const deleteSkillHandler = (e) => {
-    const btn = e.currentTarget;
+  const deleteSkillHandler = (btn) => {
     const { id } = btn.dataset;
     setSkillsData({
       ...skillsData,
@@ -55,7 +56,11 @@ export default function Skills({ isEditing, validateForm }) {
         <h2>Skills</h2>
         {isEditing ? (
           <>
-            <form noValidate={true} onSubmit={validateForm(() => addSkillHandler())}>
+            <form
+              className="no-validate-all"
+              noValidate={true}
+              onSubmit={(e) => validateForm(e, addSkillHandler)}
+            >
               <ul>
                 <FormItem
                   id="skill"
@@ -69,7 +74,7 @@ export default function Skills({ isEditing, validateForm }) {
                 <Button type="submit" text="Add"></Button>
               </ul>
             </form>
-            <form>
+            <form noValidate={true} onSubmit={(e) => validateForm(e)}>
               <ul>
                 {skillsData.skills.map((skill) => {
                   return (
@@ -78,7 +83,7 @@ export default function Skills({ isEditing, validateForm }) {
                       id={`skill_${skill.id}`}
                       value={skill.value}
                       type="text"
-                      name="skills"
+                      name="skill"
                       onChange={onChangeHandler}
                       dataAttributes={{ 'data-id': skill.id, 'data-key': 'skill' }}
                       placeholder="Edit or delete skills"
@@ -86,7 +91,13 @@ export default function Skills({ isEditing, validateForm }) {
                     >
                       <Button
                         text="Delete skill"
-                        onClick={deleteSkillHandler}
+                        onClick={(e) => {
+                          deleteRef.current = {
+                            callback: deleteSkillHandler,
+                            btn: e.currentTarget,
+                          };
+                          setModal(true);
+                        }}
                         dataAttributes={{ 'data-id': skill.id }}
                       ></Button>
                     </FormItem>